Develop

+++ src/main/java/com/chwihae/domain/bookmark/BookmarkRepository.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 @Repository
 public interface BookmarkRepository extends JpaRepository<BookmarkEntity, Long> {
 
-    long countByQuestionEntityId(Long questionId);
+    int countByQuestionEntityId(Long questionId);
 
     boolean existsByQuestionEntityIdAndUserEntityId(Long questionId, Long userId);
